[api] Add explicit JsonEncodable.Auto
Note this requires adding the following definition of exportEncoder to
any class extending Auto. Would love to know how to automate this more.

N.b. Replace MyCaseClass and arguments with your own class. If this is
used as part of an enum (trait and case classes), the trait should
extend JsonEncodable, and the case classes should extend
JsonEncodable.Auto/JsonEncodable.Object/JsonEncodable.Partial

    final case class MyCaseClass(a:A, b: B) extends JsonEncodable.Auto {
      protected def encoder: ExportedEncoder[this.type] = exportEncoder[MyCaseClass]
    }